requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
openai==1.35.10
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
schedule==1.2.0
requests==2.31.0
redis==4.6.0
cachetools==5.3.2
slowapi==0.1.9
pydantic==2.5.0

